Restore the Triton/SM use_syrk guard; class-level split_qkv reject; align update rounding
Review follow-ups on the reparent. The reparent commit claimed
_resolve_use_syrk duplicated a parent-side check - that was wrong: the
Triton >= 3.4.0 + SM-allowlist validation lives in emerging-optimizers'
Muon.__init__, which LayerShardedMuon no longer inherits, and
TensorParallelMuon only version-gates use_syrk. Without the guard,
Triton < 3.4.0 asserted on the first optimizer step instead of downgrading
at construction, and unvalidated SM architectures ran the SYRK kernel
silently. Restore _resolve_use_syrk and resolve BEFORE the parent call, so
unfit hardware downgrades ahead of the parent's EO-version raise and the
parent's closure captures the resolved value (fallback and layer-sharded
paths agree). The 'auto-disabled when Triton/SM requirements are unmet'
documentation is accurate again. TensorParallelMuon itself has the same
gap (pre-existing, all SMs accepted); moving the guard into it is left as
a separate change.

Reject split_qkv=True in LayerShardedMuon.__init__: the pass-through made
the parent's fallback/degenerate paths split QKV while the layer-sharded
path does not, so the update rule of one optimizer would depend on whether
param_ns_homes happened to be set. validate_args keeps its assert for the
wired path; the class-level ValueError covers direct construction.

Drop the dtype cast in _apply_update: the empty-homes fallback is a fifth
update path through OrthogonalizedOptimizer.step's uncast
p.add_(orth_grad, alpha=-lr), which computes the fused multiply-add in
promoted precision and downcasts once on store. Casting first gave bf16
params different rounding on the layer-sharded paths than on the fallback
and than duplicated-mode TensorParallelMuon; now all five paths round
identically (no-op for Megatron's fp32 main params).

Also update the stale '--muon-tp-mode is ignored' warning: tp_mode now
reaches the constructor and applies on the delegated fallback/degenerate
paths; warn (for any non-default value) that it does not affect the
layer-sharded exchange itself. Tests: restore the hardware-gate bypass in
the batched-syrk derivation test, mirror the uncast add in the
degenerate-syrk reference, and add constructor-guard tests for the
hardware downgrade and the split_qkv reject.

+++ tests/unit_tests/optimizer/test_layer_sharded_muon.py
@@ -1444,11 +1444,11 @@ def test_degenerate_domain_honours_use_syrk():
     with _prec_ctx("medium"):
         orth = newton_schulz(mom.float(), steps=5, coefficient_type="quintic", use_syrk=True)
     scale = max(full.shape) ** 0.5
-    # Mirror _apply_update's exact op: cast, then a single fused add_(alpha=-lr).
-    # A separate `full - lr * x` rounds the lr multiply first and is NOT bitwise
-    # equal to add_ with alpha.
+    # Mirror _apply_update's exact op: a single fused add_(alpha=-lr), no dtype
+    # cast (matching the base class's fallback-path update). A separate
+    # `full - lr * x` rounds the lr multiply first and is NOT bitwise equal.
     ref = full.clone()
-    ref.add_((orth * scale * 1.0).to(ref.dtype), alpha=-lr)
+    ref.add_(orth * scale * 1.0, alpha=-lr)
     assert torch.equal(p.data, ref), (
         f"Degenerate-domain use_syrk mismatch: max_diff={(p.data - ref).abs().max().item():.2e}"
     )
